comp_matrices have now floats instead of ints

+++ src/AHP/AhpBody.py
@@ -26,6 +26,7 @@ class AhpBody:
         return question_list
 
     def createMatrices(self):
-        self.comp_matrices = [np.matrix([[0]*(len(self.planet_names))]*(len(self.planet_names))) for _ in range(len(self.criterions))]
-        self.criterion_importance_m = np.matrix[[0]*(len(self.criterions)**2)]
+        self.comp_matrices = [np.matrix([[0]*(len(self.planet_names))]*(len(self.planet_names)), dtype=float)
+                              for _ in range(len(self.criterions))]
+        self.criterion_importance_m = np.matrix([[0]*(len(self.criterions)**2)],dtype=float)
